ai-service/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
pydantic==2.5.0
requests==2.31.0
python-dotenv==1.0.0
PyPDF2==3.0.1
python-docx==1.1.0
Pillow==10.1.0
nltk==3.8.1
scikit-learn==1.3.2
numpy==1.24.3
pandas==2.1.4
groq==0.4.1
google-generativeai==0.3.2
langchain==0.1.0
langchain-community==0.0.10
spacy==3.7.2
textstat==0.7.3